+++ spec/features/character_spec.rb
@@ -17,32 +17,31 @@ describe "Get character's public information" do
     specify { expect(subject.not_modified?).to eq(false) }
 
     specify do
-      expect(subject.as_json).to eq(corporation_id: 98_134_807,
+      expect(subject.as_json).to eq(corporation_id: 98_565_696,
         birthday: "2010-01-15T15:26:00Z",
         name: "Johnn Dillinger",
         gender: "male",
         race_id: 2,
         bloodline_id: 4,
         description: "",
-        alliance_id: 99_005_443,
-        ancestry_id: 24,
-        security_status: 1.8694881661345457,
+        alliance_id: 99_008_595,
+        security_status: 3.83155339,
         faction_id: nil,
         title: nil)
     end
 
-    specify { expect(subject.etag).to eq("22c39689783a86032b8d43fa0b2e8f4809c4f38a585e39471035aa8b") }
+    specify { expect(subject.etag).to eq("5e57d726eda270ba0e8542bdfbca874c844e5cdbf4b5599eeca79d1f") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(34) }
+    specify { expect(subject.error_limit_reset).to eq(47) }
   end
 
   context "when etag is set" do
     let(:options) do
       {
         character_id: 1_337_512_245,
-        etag: "22c39689783a86032b8d43fa0b2e8f4809c4f38a585e39471035aa8b"
+        etag: "5e57d726eda270ba0e8542bdfbca874c844e5cdbf4b5599eeca79d1f"
       }
     end
 
@@ -54,10 +53,10 @@ describe "Get character's public information" do
 
     specify { expect(subject.not_modified?).to eq(true) }
 
-    specify { expect(subject.etag).to eq("22c39689783a86032b8d43fa0b2e8f4809c4f38a585e39471035aa8b") }
+    specify { expect(subject.etag).to eq("5e57d726eda270ba0e8542bdfbca874c844e5cdbf4b5599eeca79d1f") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(28) }
+    specify { expect(subject.error_limit_reset).to eq(7) }
   end
 end